+++ apps/web/app/review-modal/components/media-info.tsx
@@ -19,21 +19,21 @@ interface MediaInfoProps {
   date: string;
   likeCount: number;
   isLiked?: boolean;
+  userStatus: boolean;
 }
 
 export default function MediaInfo({
   reviewId,
   user,
   date,
-  likeCount,
+  likeCount: initialLikeCount,
   isLiked: initialIsLiked = false,
+  userStatus,
 }: MediaInfoProps) {
-  const [isLiked, setIsLiked] = useState(initialIsLiked);
-  const likeToggle = useReviewLikeToggle();
-
-  useEffect(() => {
-    setIsLiked(initialIsLiked);
-  }, [initialIsLiked]);
+  const { likeMutation, isLiked, likeCount } = useReviewLikeToggle(
+    initialIsLiked,
+    initialLikeCount
+  );
 
   return (
     <div className="absolute bottom-0 left-0 z-10 flex h-[16rem] w-full items-center justify-between rounded-bl-xl bg-gradient-to-t from-black/60 to-transparent p-[1.6rem]">
@@ -49,12 +49,10 @@ export default function MediaInfo({
         <ReactionToggle
           variant="vertical"
           pressed={isLiked}
-          onPressedChange={(pressed) => {
-            setIsLiked(pressed);
-            likeToggle.mutate(reviewId);
+          onPressedChange={() => {
+            likeMutation.mutate(reviewId);
           }}
           className="text-white"
-          disabled={likeToggle.isPending}
         >
           {isLiked ? <SvgGoodFill /> : <SvgGoodOutline />}
           <span className="text-en-body1">{likeCount}</span>
