+++ apps/web/app/review-modal/components/review-modal-layout.tsx
@@ -27,6 +27,7 @@ export interface ReviewModallayoutProps {
   productImageUrl: string;
   onClose?: () => void;
   children?: React.ReactNode;
+  userStatus: boolean;
 }
 
 export default function ReviewModallayout({
@@ -45,6 +46,7 @@ export default function ReviewModallayout({
   productImageUrl,
   isReceipt,
   onClose,
+  userStatus,
 }: ReviewModallayoutProps) {
   return (
     <div className="relative flex h-full w-full items-center justify-center">
@@ -53,6 +55,7 @@ export default function ReviewModallayout({
         <div className="relative flex w-[55.2rem] flex-col">
           <MediaViewer mediaList={mediaList} />
           <MediaInfo
+            userStatus={userStatus}
             reviewId={id}
             user={user}
             date={user.uploadAt}
